--- src/mock/filter.js
@@ -1,15 +1,0 @@
-import {filter} from '../utils/filter.js';
-
-function generateFilter(points) {
-  return Object.entries(filter).map(
-    ([filterType, filterPoints]) => ({
-      type: filterType,
-      count: filterPoints(points).length,
-      tripPoints: filterPoints(points)
-
-    })
-  );
-}
-
-export {generateFilter};
-
